+++ packages/eas-cli/src/commands/update/index.ts
@@ -200,7 +200,7 @@ export default class UpdatePublish extends EasCommand {
 
     await maybeWarnAboutEasOutagesAsync(graphqlClient, [StatuspageServiceName.EasUpdate]);
 
-    await ensureEASUpdatesIsConfiguredAsync(graphqlClient, {
+    await ensureEASUpdateIsConfiguredAsync(graphqlClient, {
       exp: expPossiblyWithoutEasUpdateConfigured,
       platform: getRequestedPlatform(platformFlag),
       projectDir,